ft_connect: Rename pe_connect app
We will need additional work on the database to make it work
(content_types, db table name, saved migrations)

This will come in another commit
