+++ itou/openid_connect/ft_connect/apps.py
@@ -1,0 +1,6 @@
+from django.apps import AppConfig
+
+
+class FranceTravailConnectConfig(AppConfig):
+    name = "itou.openid_connect.ft_connect"
+    verbose_name = "FT Connect"
