+++ itou/openid_connect/ft_connect/constants.py
