+++ itou/openid_connect/management/commands/cleanup_oidc_states.py
@@ -1,7 +1,7 @@
 from django.template.defaultfilters import pluralize
 
 from itou.openid_connect.france_connect.models import FranceConnectState
-from itou.openid_connect.pe_connect.models import PoleEmploiConnectState
+from itou.openid_connect.ft_connect.models import PoleEmploiConnectState
 from itou.openid_connect.pro_connect.models import ProConnectState
 from itou.utils.command import BaseCommand
