--- itou/openid_connect/pe_connect/apps.py
@@ -1,6 +1,0 @@
-from django.apps import AppConfig
-
-
-class PoleEmploiConnectConfig(AppConfig):
-    name = "itou.openid_connect.pe_connect"
-    verbose_name = "PE Connect"
