+++ itou/templates/signup/includes/peamu_button.html
@@ -1,7 +1,7 @@
 {% load matomo %}
 {% load static %}
 
-<form method="post" action="{% url 'pe_connect:authorize' %}">
+<form method="post" action="{% url 'ft_connect:authorize' %}">
     {% csrf_token %}
     <input type="image" src="{% static 'img/peamu_btn.svg' %}" alt="Se connecter avec votre compte France Travail" {% matomo_event "inscription-candidat" "clic" "se-connecter-avec-pe" %}>
 </form>
