+++ tests/openid_connect/test_cleanup_oidc_states.py
@@ -6,7 +6,7 @@ from pytest_django.asserts import assertQuerySetEqual
 
 from itou.openid_connect.constants import OIDC_STATE_CLEANUP
 from itou.openid_connect.france_connect.models import FranceConnectState
-from itou.openid_connect.pe_connect.models import PoleEmploiConnectState
+from itou.openid_connect.ft_connect.models import PoleEmploiConnectState
 from itou.openid_connect.pro_connect.models import ProConnectState
